[Refactor][GroupedGemm] Replace 3WG kernel with V2 implementation
V2 (cooperative + pingpong) is strictly superior to the old 3WG kernel
(20-26% faster on prefill). This commit:

- Renames grouped_gemm_persistent_v2.py → grouped_gemm_persistent_3wg.py
- Renames class GroupedGemmPersistentV2Kernel → GroupedGemmPersistent3WGKernel
- Removes the old 3WG implementation (superseded)
- Removes bench_grouped_gemm_v2_vs_3wg.py (comparing same kernel)
- Updates __init__.py and test file accordingly

Interface unchanged: same __init__ and forward() signatures.
Default config: bm=128 bn=256 ns=3 (cooperative, autotuned best).
