+++ tileops/kernels/grouped_gemm/__init__.py
@@ -1,12 +1,10 @@
 from .grouped_gemm import _DEFAULT_CONFIGS, GroupedGemmKernel
 from .grouped_gemm_persistent import GroupedGemmPersistentKernel
 from .grouped_gemm_persistent_3wg import GroupedGemmPersistent3WGKernel
-from .grouped_gemm_persistent_v2 import GroupedGemmPersistentV2Kernel
 
 __all__ = [
     "_DEFAULT_CONFIGS",
     "GroupedGemmKernel",
     "GroupedGemmPersistent3WGKernel",
     "GroupedGemmPersistentKernel",
-    "GroupedGemmPersistentV2Kernel",
 ]
